+++ src/main/java/duke/Parser.java
@@ -1,6 +1,11 @@
 package duke;
-
-import duke.commands.*;
+import duke.commands.Command;
+import duke.commands.ListCommand;
+import duke.commands.ExitCommand;
+import duke.commands.MarkCommand;
+import duke.commands.DeleteCommand;
+import duke.commands.AddCommand;
+import duke.commands.DefaultCommand;
 
 public class Parser {
