--- src/main/java/duke/events/Events.java
@@ -1,23 +1,0 @@
-package duke.events;
-
-public class Events extends Task {
-
-    private String taskName;
-    private String time;
-    private boolean isDone = false;
-
-    public Events(String taskName, String time) {
-        super(taskName);
-        this.time = time;
-    }
-
-    @Override
-    public String toString() {
-        return "[E] " + super.toString() + "(at:" + time + ")";
-    }
-
-    @Override
-    public String getSaveData() {
-        return " E" + " | " + super.isDone() + " | " + taskName + " | " + time;
-    }
-}
