--- src/main/java/duke/events/ToDos.java
@@ -1,21 +1,0 @@
-package duke.events;
-
-public class ToDos extends Task {
-
-    private String taskName;
-
-    public ToDos(String taskName) {
-        super(taskName);
-        this.taskName = taskName;
-    }
-
-    @Override
-    public String toString() {
-        return "[T] " + super.toString();
-    }
-
-    @Override
-    public String getSaveData() {
-        return "T" + " | " + super.isDone() + " | " + taskName;
-    }
-}
