+++ src/test/java/duke/ToDoTest.java
@@ -1,5 +1,5 @@
 package duke;
-import duke.events.ToDos;
+import duke.events.ToDo;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -7,7 +7,7 @@ public class ToDoTest {
 
     @Test
     public void todoSaveData(){
-        ToDos t = new ToDos("Buy milk");
+        ToDo t = new ToDo("Buy milk");
         String expected = "T | false | Buy milk";
         String actual = t.getSaveData();
         assertEquals(expected, actual);
@@ -15,7 +15,7 @@ public class ToDoTest {
 
     @Test
     public void todoToString(){
-        ToDos t = new ToDos("Buy milk");
+        ToDo t = new ToDo("Buy milk");
         String expected = "[T] [ ] Buy milk";
         String actual = t.toString();
         assertEquals(expected, actual);
